requirements/base.txt
Flask==2.2.2
flask-marshmallow==0.14.0
redis==4.3.2
gunicorn==20.1.0
python-dotenv==0.20.0
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
celery==5.2.7
xmltodict==0.13.0
